add extendSelect method,add sequence of sql

+++ src/main/java/com/xw/supercar/entity/InPartInfo.java
@@ -1,5 +1,7 @@
 package com.xw.supercar.entity;
 
+import java.math.BigDecimal;
+
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -20,7 +22,7 @@ public class InPartInfo extends BaseDateEntity {
     private Integer count;
 
     /** 进货价 */
-    private Long cost;
+    private BigDecimal cost;
 
     /** 供应商，数据字典外键 */
     private String supplierLK;
